+++ api/src/certification/configuration/application/complementary-certification-route.js
@@ -1,3 +1,5 @@
+import Joi from 'joi';
+
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { complementaryCertificationController } from './complementary-certification-controller.js';
 
@@ -27,8 +29,36 @@ const register = async function (server) {
         ],
       },
     },
+    {
+      method: 'GET',
+      path: '/api/admin/complementary-certifications/attachable-target-profiles',
+      config: {
+        validate: {
+          query: Joi.object({
+            searchTerm: Joi.string().allow(null, '').optional(),
+          }),
+        },
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.hasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        handler: complementaryCertificationController.searchAttachableTargetProfilesForComplementaryCertifications,
+        tags: ['api', 'admin'],
+        notes: [
+          'Cette route est restreinte aux utilisateurs authentifiés avec le rôle Super Admin, Support et Métier',
+          'Elle renvoie les profils cibles qui peuvent être attachés à un certification complémentaire.',
+        ],
+      },
+    },
   ]);
 };
 
-const name = 'configuration-complementary-certifications-api';
+const name = 'complementary-certifications-api';
 export { name, register };
